+++ cmd/epp/main.go
@@ -160,7 +160,7 @@ func run() error {
 	datastore := datastore.NewDatastore(ctx, pmf)
 
 	// Set up mapper for metric scraping.
-	mapping, err := backend.NewMetricMapping(
+	mapping, err := vllm.NewMetricMapping(
 		*totalQueuedRequestMetric,
 		*kVCacheUsageMetric,
 		*loraRequestInfoMetric,
@@ -169,7 +169,7 @@ func run() error {
 		setupLog.Error(err, "Failed to create metric mapping from flags.")
 		return err
 	}
-	provider := backend.NewProvider(&backend.PodMetricsClientImpl{MetricMapping: mapping}, datastore)
+	provider := backend.NewProvider(&vllm.PodMetricsClientImpl{MetricMapping: mapping}, datastore)
 	//
 	serverRunner := &runserver.ExtProcServerRunner{
 		GrpcPort:                                 *grpcPort,
