+++ pkg/epp/backend/vllm/metrics.go
@@ -14,7 +14,6 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-// Package vllm provides vllm specific pod metrics implementation.
 package vllm
 
 import (
@@ -30,60 +29,49 @@ import (
 	"github.com/prometheus/common/expfmt"
 	"go.uber.org/multierr"
 	"sigs.k8s.io/controller-runtime/pkg/log"
-	"sigs.k8s.io/gateway-api-inference-extension/pkg/epp/backend/metrics"
+	"sigs.k8s.io/gateway-api-inference-extension/pkg/epp/datastore"
 	logutil "sigs.k8s.io/gateway-api-inference-extension/pkg/epp/util/logging"
 )
 
-// Metric names used in the vLLM metrics implementation.
-// Refer to the protocol doc for more details:
-// https://github.com/kubernetes-sigs/gateway-api-inference-extension/tree/main/docs/proposals/003-model-server-protocol
 const (
-	LoraRequestInfoMetricName                = "vllm:lora_requests_info"
+	// Hardcoded vLLM specific LoRA metrics
 	LoraRequestInfoRunningAdaptersMetricName = "running_lora_adapters"
 	LoraRequestInfoWaitingAdaptersMetricName = "waiting_lora_adapters"
 	LoraRequestInfoMaxAdaptersMetricName     = "max_lora"
-	// TODO: Replace these with the num_tokens_running/waiting below once we add those to the fork.
-	RunningQueueSizeMetricName = "vllm:num_requests_running"
-	WaitingQueueSizeMetricName = "vllm:num_requests_waiting"
-	/* TODO: Uncomment this once the following are added to the fork.
-	RunningQueueSizeMetricName        = "vllm:num_tokens_running"
-	WaitingQueueSizeMetricName        = "vllm:num_tokens_waiting"
-	*/
-	KVCacheUsagePercentMetricName = "vllm:gpu_cache_usage_perc"
 )
 
-type PodMetricsClientImpl struct{}
+type PodMetricsClientImpl struct {
+	MetricMapping *MetricMapping
+}
 
 // FetchMetrics fetches metrics from a given pod.
 func (p *PodMetricsClientImpl) FetchMetrics(
 	ctx context.Context,
-	pod *metrics.Pod,
-	existing *metrics.Metrics,
+	existing *datastore.PodMetrics,
 	port int32,
-) (*metrics.Metrics, error) {
-	logger := log.FromContext(ctx).V(logutil.TRACE)
+) (*datastore.PodMetrics, error) {
+	logger := log.FromContext(ctx)
+	loggerDefault := logger.V(logutil.DEFAULT)
 
-	// Currently the metrics endpoint is hard-coded, which works with vLLM.
-	// TODO(https://github.com/kubernetes-sigs/gateway-api-inference-extension/issues/16): Consume this from InferencePool config.
-	url := "http://" + pod.Address + ":" + strconv.Itoa(int(port)) + "/metrics"
+	url := "http://" + existing.Address + ":" + strconv.Itoa(int(port)) + "/metrics"
 
 	req, err := http.NewRequestWithContext(ctx, http.MethodGet, url, nil)
 	if err != nil {
-		logger.Error(err, "Failed create HTTP request", "method", http.MethodGet, "url", url)
+		loggerDefault.Error(err, "Failed create HTTP request", "method", http.MethodGet, "url", url)
 		return nil, fmt.Errorf("failed to create request: %v", err)
 	}
 	resp, err := http.DefaultClient.Do(req)
 	if err != nil {
-		logger.Error(err, "Failed to fetch metrics", "pod", pod.NamespacedName)
-		return nil, fmt.Errorf("failed to fetch metrics from %s: %w", pod.NamespacedName, err)
+		loggerDefault.Error(err, "Failed to fetch metrics", "pod", existing.NamespacedName)
+		return nil, fmt.Errorf("failed to fetch metrics from %s: %w", existing.NamespacedName, err)
 	}
 	defer func() {
 		_ = resp.Body.Close()
 	}()
 
 	if resp.StatusCode != http.StatusOK {
-		logger.Error(nil, "Unexpected status code returned", "pod", pod.NamespacedName, "statusCode", resp.StatusCode)
-		return nil, fmt.Errorf("unexpected status code from %s: %v", pod.NamespacedName, resp.StatusCode)
+		loggerDefault.Error(nil, "Unexpected status code returned", "pod", existing.NamespacedName, "statusCode", resp.StatusCode)
+		return nil, fmt.Errorf("unexpected status code from %s: %v", existing.NamespacedName, resp.StatusCode)
 	}
 
 	parser := expfmt.TextParser{}
@@ -91,74 +79,70 @@ func (p *PodMetricsClientImpl) FetchMetrics(
 	if err != nil {
 		return nil, err
 	}
-	return promToPodMetrics(logger, metricFamilies, existing)
+	return p.promToPodMetrics(logger, metricFamilies, existing)
 }
 
-// promToPodMetrics updates internal pod metrics with scraped prometheus metrics.
-// A combined error is returned if errors occur in one or more metric processing.
-// it returns a new PodMetrics pointer which can be used to atomically update the pod metrics map.
-func promToPodMetrics(
+// promToPodMetrics updates internal pod metrics with scraped Prometheus metrics.
+func (p *PodMetricsClientImpl) promToPodMetrics(
 	logger logr.Logger,
 	metricFamilies map[string]*dto.MetricFamily,
-	existing *metrics.Metrics,
-) (*metrics.Metrics, error) {
+	existing *datastore.PodMetrics,
+) (*datastore.PodMetrics, error) {
 	var errs error
 	updated := existing.Clone()
-	runningQueueSize, err := getLatestMetric(logger, metricFamilies, RunningQueueSizeMetricName)
-	errs = multierr.Append(errs, err)
-	if err == nil {
-		updated.RunningQueueSize = int(runningQueueSize.GetGauge().GetValue())
-	}
-	waitingQueueSize, err := getLatestMetric(logger, metricFamilies, WaitingQueueSizeMetricName)
-	errs = multierr.Append(errs, err)
-	if err == nil {
-		updated.WaitingQueueSize = int(waitingQueueSize.GetGauge().GetValue())
-	}
-	cachePercent, err := getLatestMetric(logger, metricFamilies, KVCacheUsagePercentMetricName)
-	errs = multierr.Append(errs, err)
-	if err == nil {
-		updated.KVCacheUsagePercent = cachePercent.GetGauge().GetValue()
+
+	if p.MetricMapping.TotalQueuedRequests != nil {
+		queued, err := p.getMetric(logger, metricFamilies, *p.MetricMapping.TotalQueuedRequests)
+		if err == nil {
+			updated.WaitingQueueSize = int(queued.GetGauge().GetValue())
+		} else {
+			errs = multierr.Append(errs, err)
+		}
 	}
 
-	loraMetrics, _, err := getLatestLoraMetric(logger, metricFamilies)
-	errs = multierr.Append(errs, err)
-	/* TODO: uncomment once this is available in vllm.
-	kvCap, _, err := getGaugeLatestValue(metricFamilies, KvCacheMaxTokenCapacityMetricName)
-	errs = multierr.Append(errs, err)
-	if err != nil {
-		updated.KvCacheMaxTokenCapacity = int(kvCap)
+	if p.MetricMapping.KVCacheUtilization != nil {
+		usage, err := p.getMetric(logger, metricFamilies, *p.MetricMapping.KVCacheUtilization)
+		if err == nil {
+			updated.KVCacheUsagePercent = usage.GetGauge().GetValue()
+		} else {
+			errs = multierr.Append(errs, err)
+		}
 	}
-	*/
 
-	if loraMetrics != nil {
-		updated.ActiveModels = make(map[string]int)
-		for _, label := range loraMetrics.GetLabel() {
-			if label.GetName() == LoraRequestInfoRunningAdaptersMetricName {
-				if label.GetValue() != "" {
-					adapterList := strings.Split(label.GetValue(), ",")
-					for _, adapter := range adapterList {
-						updated.ActiveModels[adapter] = 0
+	// Handle LoRA metrics (only if all LoRA MetricSpecs are present)
+	if p.MetricMapping.LoraRequestInfo != nil {
+		loraMetrics, _, err := p.getLatestLoraMetric(logger, metricFamilies)
+		errs = multierr.Append(errs, err)
+
+		if loraMetrics != nil {
+			updated.ActiveModels = make(map[string]int)
+			for _, label := range loraMetrics.GetLabel() {
+				if label.GetName() == LoraRequestInfoRunningAdaptersMetricName {
+					if label.GetValue() != "" {
+						adapterList := strings.Split(label.GetValue(), ",")
+						for _, adapter := range adapterList {
+							updated.ActiveModels[adapter] = 0
+						}
 					}
 				}
-			}
-			if label.GetName() == LoraRequestInfoWaitingAdaptersMetricName {
-				if label.GetValue() != "" {
-					adapterList := strings.Split(label.GetValue(), ",")
-					for _, adapter := range adapterList {
-						updated.ActiveModels[adapter] = 0
+				if label.GetName() == LoraRequestInfoWaitingAdaptersMetricName {
+					if label.GetValue() != "" {
+						adapterList := strings.Split(label.GetValue(), ",")
+						for _, adapter := range adapterList {
+							updated.ActiveModels[adapter] = 0
+						}
 					}
 				}
-			}
-			if label.GetName() == LoraRequestInfoMaxAdaptersMetricName {
-				if label.GetValue() != "" {
-					updated.MaxActiveModels, err = strconv.Atoi(label.GetValue())
-					if err != nil {
-						errs = multierr.Append(errs, err)
+				if label.GetName() == LoraRequestInfoMaxAdaptersMetricName {
+					if label.GetValue() != "" {
+						updated.MaxActiveModels, err = strconv.Atoi(label.GetValue())
+						if err != nil {
+							errs = multierr.Append(errs, err)
+						}
 					}
 				}
 			}
 		}
-
 	}
 
 	return updated, errs
@@ -168,62 +152,80 @@ func promToPodMetrics(
 // reason its specially fetched is because each label key value pair permutation generates new series
 // and only most recent is useful. The value of each series is the creation timestamp so we can
 // retrieve the latest by sorting the value.
-func getLatestLoraMetric(logger logr.Logger, metricFamilies map[string]*dto.MetricFamily) (*dto.Metric, time.Time, error) {
-	loraRequests, ok := metricFamilies[LoraRequestInfoMetricName]
+func (p *PodMetricsClientImpl) getLatestLoraMetric(logger logr.Logger, metricFamilies map[string]*dto.MetricFamily) (*dto.Metric, time.Time, error) {
+	if p.MetricMapping.LoraRequestInfo == nil {
+		return nil, time.Time{}, nil // No LoRA metrics configured
+	}
+
+	loraRequests, ok := metricFamilies[p.MetricMapping.LoraRequestInfo.MetricName]
 	if !ok {
-		logger.V(logutil.TRACE).Error(nil, "Metric family not found", "name", LoraRequestInfoMetricName)
-		return nil, time.Time{}, fmt.Errorf("metric family %q not found", LoraRequestInfoMetricName)
+		logger.V(logutil.DEFAULT).Error(nil, "Metric family not found", "name", p.MetricMapping.LoraRequestInfo.MetricName)
+		return nil, time.Time{}, fmt.Errorf("metric family %q not found", p.MetricMapping.LoraRequestInfo.MetricName)
 	}
 
 	var latest *dto.Metric
-	var latestTs float64
+	var latestTs float64 // Use float64, as Gauge.Value is float64
 
 	// Iterate over all metrics in the family.
 	for _, m := range loraRequests.GetMetric() {
-		var running, waiting string
-		// Read the label values for running and waiting adapters.
+		running := ""
+		waiting := ""
+		// Check if the metric has the expected LoRA labels.  This is important!
+		hasRequiredLabels := false
 		for _, lp := range m.GetLabel() {
 			switch lp.GetName() {
 			case LoraRequestInfoRunningAdaptersMetricName:
 				running = lp.GetValue()
+				hasRequiredLabels = true
 			case LoraRequestInfoWaitingAdaptersMetricName:
 				waiting = lp.GetValue()
+				hasRequiredLabels = true
 			}
 		}
-
-		// Ignore metrics with both labels empty. This happens when there are no running or waiting requests on
-		// the server, in this case it is best to use the last set of active adapters.
+		//Skip if it does not have the lora labels
+		if !hasRequiredLabels {
+			continue
+		}
+		// Ignore metrics with both labels empty.
 		if running == "" && waiting == "" {
 			continue
 		}
 
-		// Select the metric with the latest creation timestamp.
+		// Select the metric with the *largest Gauge Value* (which represents the timestamp).
 		if m.GetGauge().GetValue() > latestTs {
 			latestTs = m.GetGauge().GetValue()
 			latest = m
 		}
 	}
-
 	if latest == nil {
-		logger.V(logutil.TRACE).Info("Metric value Empty", "value", latest, "metric", LoraRequestInfoMetricName)
+		logger.V(logutil.TRACE).Info("Metric value Empty", "value", latest, "metric", p.MetricMapping.LoraRequestInfo.MetricName)
 		return nil, time.Time{}, nil
 	}
 
 	// Convert the gauge value (creation timestamp) to time.Time.
-	return latest, time.Unix(0, int64(latestTs*1000)), nil
+	return latest, time.Unix(0, int64(latestTs*1e9)), nil // Convert nanoseconds to time.Time
 }
 
-// getLatestMetric gets the latest metric of a family. This should be used to get the latest Gauge metric.
-// Since vllm doesn't set the timestamp in metric, this metric essentially gets the first metric.
-func getLatestMetric(logger logr.Logger, metricFamilies map[string]*dto.MetricFamily, metricName string) (*dto.Metric, error) {
-	mf, ok := metricFamilies[metricName]
+// getMetric retrieves a specific metric based on MetricSpec.
+func (p *PodMetricsClientImpl) getMetric(logger logr.Logger, metricFamilies map[string]*dto.MetricFamily, spec MetricSpec) (*dto.Metric, error) {
+	mf, ok := metricFamilies[spec.MetricName]
 	if !ok {
-		logger.V(logutil.TRACE).Error(nil, "Metric family not found", "name", metricName)
-		return nil, fmt.Errorf("metric family %q not found", metricName)
+		logger.V(logutil.DEFAULT).Error(nil, "Metric family not found", "name", spec.MetricName)
+		return nil, fmt.Errorf("metric family %q not found", spec.MetricName)
 	}
+
 	if len(mf.GetMetric()) == 0 {
-		return nil, fmt.Errorf("no metrics available for %q", metricName)
+		return nil, fmt.Errorf("no metrics available for %q", spec.MetricName)
 	}
+	// if there is a specified label, return only that metric in the family
+	if spec.Labels != nil {
+		return getLabeledMetric(logger, mf, spec)
+	}
+	return getLatestMetric(logger, mf)
+}
+
+// getLatestMetric gets the latest metric of a family (for metrics without labels).
+func getLatestMetric(logger logr.Logger, mf *dto.MetricFamily) (*dto.Metric, error) {
 	var latestTs int64
 	var latest *dto.Metric
 	for _, m := range mf.GetMetric() {
@@ -232,6 +234,54 @@ func getLatestMetric(logger logr.Logger, metricFamilies map[string]*dto.MetricFa
 			latest = m
 		}
 	}
-	logger.V(logutil.TRACE).Info("Metric value selected", "value", latest, "metric", metricName)
+
+	if latest == nil {
+		return nil, fmt.Errorf("no metrics found for %q", mf.GetName())
+	}
+
+	logger.V(logutil.TRACE).Info("Latest metric value selected", "value", latest, "metric", mf.GetName())
 	return latest, nil
+}
+
+// getLabeledMetric gets the latest metric with matching labels.
+func getLabeledMetric(logger logr.Logger, mf *dto.MetricFamily, spec MetricSpec) (*dto.Metric, error) {
+	var latestMetric *dto.Metric
+	var latestTimestamp int64 = -1 // Initialize to -1 so any timestamp is greater
+
+	for _, m := range mf.GetMetric() {
+		if labelsMatch(m.GetLabel(), spec.Labels) {
+			if m.GetTimestampMs() > latestTimestamp {
+				latestTimestamp = m.GetTimestampMs()
+				latestMetric = m
+			}
+		}
+	}
+
+	if latestMetric != nil {
+		logger.V(logutil.TRACE).Info("Labeled metric found", "value", latestMetric, "metric", spec.MetricName)
+		return latestMetric, nil
+	}
+
+	return nil, fmt.Errorf("no matching labeled metric found for %q with labels %v", spec.MetricName, spec.Labels)
+}
+
+// labelsMatch checks if a metric's labels contain all the labels in the spec.
+func labelsMatch(metricLabels []*dto.LabelPair, specLabels map[string]string) bool {
+	if len(specLabels) == 0 {
+		return true // No specific labels required
+	}
+
+	for specName, specValue := range specLabels {
+		found := false
+		for _, label := range metricLabels {
+			if label.GetName() == specName && label.GetValue() == specValue {
+				found = true
+				break
+			}
+		}
+		if !found {
+			return false // A required label is missing
+		}
+	}
+	return true // All required labels are present
 }
