+++ pkg/epp/backend/vllm/metrics_spec.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package backend
+package vllm
 
 import (
 	"fmt"
